🏗️ Architect: Modularized apts/place/models.py
Refactored the 623-line `apts/place/models.py` "God Object" into a modular package structure. Logic is now split into:
- `conditions.py`: Environmental factors (SQM, seeing) and weather integration.
- `times.py`: Astronomical event timing (rise/set times).
- `paths.py`: Celestial body trajectory calculations and plotting.
- `imaging.py`: Imaging window optimization logic.

The `Place` class in `base.py` now inherits from these specialized mixins imported from their respective modules, adhering to the Single Responsibility Principle. All relevant unit tests passed.
